fix(crm): auto-select restaurant from tracking page + drop query after
When navigating from /admin/tracking with ?restaurant_id=N, the CRM
should auto-open that record's detail panel on the right side. This
makes the tracking ↔ CRM round-trip a one-click flow.

Implementation choice: window.location.search in a useEffect, not
useSearchParams. Next.js 16 prerender blocks useSearchParams unless
wrapped in <Suspense>; for a fully client-side page that's already
gated by the auth middleware, the manual URL parse keeps the page
rendered without the Suspense boundary.

Also strips the parameter after consuming it so a manual refresh
(F5) doesn't re-trigger the auto-select.

Companion to /admin/tracking CTA "在 CRM 中開啟 ↗" (commit 56cefc1).

+++ src/app/admin/api/contact-logs/tracking/route.ts
@@ -53,7 +53,7 @@ export async function GET(request: Request) {
       ll.latest_contact_date,
       ll.latest_status,
       ll.latest_created_at,
-      (SELECT COUNT(*) FROM contact_logs cl2 WHERE cl2.restaurant_id = r.id) AS total_logs
+      (SELECT COUNT(*)::int FROM contact_logs cl2 WHERE cl2.restaurant_id = r.id) AS total_logs
     FROM latest_logs ll
     JOIN restaurants r ON r.id = ll.restaurant_id
     WHERE r.disabled_at IS NULL
